# docs/openapi.yaml
openapi: 3.1.0
info:
  title: AIHub API
  version: 1.0.0
  description: Pay-per-use AI marketplace API with x402 settlement on Algorand.
servers:
  - url: http://localhost:8080
paths:
...
  /admin/analytics:
    get:
      summary: Admin analytics
      security:
        - bearerAuth: []
      responses:
        "200":
          description: Admin metrics
components:
  securitySchemes:
    bearerAuth:
      type: http
      scheme: bearer
      bearerFormat: JWT